add Task23 Изменил типы данных

+++ HomeworkFist/Task23/Program.cs
@@ -1,9 +1,9 @@
 ﻿// Задача 23 Напишите программу, которая принимает на вход число (N) и выдаёт таблицу кубов
 Console.Write("Ведите последнее число таблицы: ");
-double N = Convert.ToDouble(Console.ReadLine());
+int N = Convert.ToInt32(Console.ReadLine());
 Console.Write("Ведите интересующую степень: ");
-double A = Convert.ToDouble(Console.ReadLine());
-double count = 0;
+int A = Convert.ToInt32(Console.ReadLine());
+int count = 0;
 double step = 0;
 while (count <= N)
 {
